requirements.txt
pyyaml
rpi_ws281x
adafruit-circuitpython-neopixel
adafruit-blinka
